requirements.txt
Flask==3.0.3
requests==2.32.3
pandas==2.2.2
openai==1.30.1
python-dotenv==1.0.1